feat: LeafOptions for multi-leaf discrimination, named keywords, build() safety
- Add LeafOptions<K> interface to @sittir/types for discriminating leaf
  types in multi-kind field unions (nodeKind + optional text)
- Add listNamedKeywords() to grammar reader — detects leaf kinds with
  constant text from grammar.json rules (STRING, SEQ, precedence wrappers)
- Named keywords generate zero-arg factories (ir.self(), ir.null(), etc.)
  with optional text defaulting to the fixed grammar text
- Multi-leaf fields emit LeafOptions<'kind1'> | LeafOptions<'kind2'>
  instead of bare `| string`, enabling switch dispatch on nodeKind
- Fix build() null safety: non-constructor required fields use optional
  chaining to avoid runtime errors on partially-constructed builders
- Regenerate all three downstream packages (rust, typescript, python)
- Update 47 hand-written test files for zero-arg keyword calls

+++ packages/python/tests/complex-pattern.test.ts
@@ -3,7 +3,7 @@ import { ir } from '../src/builder.js';
 
 describe('complex_pattern', () => {
   it('should build with correct kind', () => {
-    const builder = ir.complexPattern(ir.float('test'), ir.float('test'));
+    const builder = ir.complexPattern(ir.integer('test'), ir.integer('test'));
     const node = builder.build();
     expect(node.kind).toBe('complex_pattern');
     expect(Array.isArray((node as any).children)).toBe(true);
@@ -12,7 +12,7 @@ describe('complex_pattern', () => {
   });
 
   it('should produce a valid CST node', () => {
-    const builder = ir.complexPattern(ir.float('test'), ir.float('test'));
+    const builder = ir.complexPattern(ir.integer('test'), ir.integer('test'));
     const cst = builder.toCST();
     expect(cst.type).toBe('complex_pattern');
     expect(cst.isNamed).toBe(true);
@@ -21,7 +21,7 @@ describe('complex_pattern', () => {
   });
 
   it('should pass fast validation', () => {
-    const builder = ir.complexPattern(ir.float('test'), ir.float('test'));
+    const builder = ir.complexPattern(ir.integer('test'), ir.integer('test'));
     expect(() => builder.render('fast')).not.toThrow();
   });
 });
